# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Be
  description: |
    The API provides CSR backend API.
basePath: /api/
schemes:
  - https
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/refresh:
    post:
      description: 'Returns new token pair for an authorized User'
      tags:
        - Users
      operationId: Refresh
      consumes:
        - "application/json"
      parameters:
        - name: 'Refresh token'
          in: 'body'
          required: true
          description: 'Refresh payload'
          schema:
            $ref: '#/definitions/RefreshToken'

      responses:
        200:
          description: Successful refresh.
          schema:
            $ref: '#/definitions/TokenPair'
        404:
          description: Refresh token not found.
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
...
  /v1/users/me:
    get:
      summary: Get the current user data.
      security:
        - Bearer: [ ]
      tags:
        - Users
      operationId: getCurrentUser
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/UserInfo"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
    patch:
      summary: Edit selected fields of the current user.
      security:
        - Bearer: [ ]
      tags:
        - Users
      operationId: patchUser
      consumes:
        - application/json
      parameters:
        - name: userPatch
          in: body
          description: The patch to apply to the user.
          required: true
          schema:
            $ref: "#/definitions/PatchUserRequest"
      responses:
        204:
          description: Success
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: Delete the current user.
      security:
        - Bearer: [ ]
      tags:
        - Users
      operationId: deleteCurrentUser
      responses:
        200:
          description: Success
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
  /v1/users/me/password:
    patch:
      summary: Change the current user password.
      security:
        - Bearer: [ ]
      tags:
        - Users
      operationId: changePassword
      consumes:
        - application/json
      parameters:
        - name: passwordPatch
          in: body
          description: The patch to apply to the user password.
          required: true
          schema:
            $ref: "#/definitions/PatchPasswordRequest"
      responses:
        204:
          description: Success
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        403:
          description: Wrong old password
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  /v1/users/me/email:
    patch:
      summary: Change the current email.
      security:
        - Bearer: [ ]
      tags:
        - Users
      operationId: changeEmail
      consumes:
        - application/json
      parameters:
        - name: emailPatch
          in: body
          description: The patch to apply to the user email.
          required: true
          schema:
            $ref: "#/definitions/PatchEmailRequest"
      responses:
        204:
          description: Success
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
...
  /v1/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: user id
        type: integer
    post:
      consumes:
        - application/json
      parameters:
        - name: UpdateUserTask
          in: body
          description: Data to update.
          required: true
          schema:
            $ref: "#/definitions/UpdateUserTask"
      tags:
        - Users
      operationId: UserUpdate
      summary: Updates user information by its id.
      security:
        - Bearer: [ ]
      responses:
        200:
          description: User information has been updated.
          schema:
            $ref: "#/definitions/SuccessUserUpdateResponse"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
    delete:
      consumes:
        - application/json
      tags:
        - Users
      operationId: deleteUser
      summary: Delete user by ID
      security:
        - Bearer: [ ]
      responses:
        204:
          description: Success
        404:
          description: User not found
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
  /v1/users/{userId}/readonly-access:
    put:
      summary: Update user's read-only access status
      parameters:
        - name: userId
          in: path
          description: ID of the user to update
          required: true
          type: integer
        - name: body
          in: body
          description: Request body containing the new read-only access status
          required: true
          schema:
            type: object
            properties:
              is_readonly:
                type: boolean
                description: New read-only access status for the user
      tags:
        - Users
      security:
        - Bearer: [ ]
      operationId: UpdateReadonlyAccess
      responses:
        204:
          description: Success
        400:
          description: Bad request
          schema:
            $ref: "#/definitions/Error"
        401:
          description: Unauthorized
          schema:
            $ref: "#/definitions/Error"
        403:
          description: Forbidden
          schema:
            $ref: "#/definitions/Error"
        404:
          description: User not found
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error.
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Error:
    type: object
    required:
      - data
    properties:
      data:
        type: object
        properties:
          correlationId:
            type: string
          message:
            type: string
...
  UpdateUserTask:
    type: object
    required:
      - data
    description: Update user object
    properties:
      data:
        type: object
        properties:
          name:
            type: string
            description: Name of user.
          surname:
            type: string
            description: Surname of user.
          patronymic:
            type: string
            description: Patronymic data.
          passport_series:
            type: string
          passport_number:
            type: string
          passport_authority:
            type: string
          passport_issue_date:
            type: string
          phone_number:
            type: string
            description: Phone number.
          email:
            type: string
          type:
            type: string
          active_areas:
            type: array
            items:
              type: integer
          org_name:
            type: string
          vk:
            type: string
  PatchUserRequest:
    type: object
    properties:
      name:
        type: string
      surname:
        type: string
      passport_series:
        type: string
      passport_number:
        type: string
      passport_authority:
        type: string
      passport_issue_date:
        type: string
        format: date
      phone:
        type: string
      org_name:
        type: string
      website:
        type: string
      vk:
        type: string
  PatchPasswordRequest:
    type: object
    properties:
        old_password:
            type: string
        new_password:
            type: string
  PatchEmailRequest:
    type: object
    properties:
        new_email:
            type: string
      # User
...
  RefreshToken:
    type: object
    required:
      - refreshToken
    properties:
      refreshToken:
        type: string
  TokenPair:
    type: object
    required:
      - accessToken
      - refreshToken
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
...
  User:
    type: object
    required:
      - id
      - createTime
      - roleId
      - login
    properties:
      id:
        type: integer
      login:
        type: string
      createTime:
        type: string
        format: date-time
      roleId:
        type: integer
      name:
        type: string
        description: Name of user.
      surname:
        type: string
        description: Surname of user.
      patronymic:
        type: string
        description: Patronymic data.
      passport_series:
        type: string
      passport_number:
        type: string
      passport_authority:
        type: string
      passport_issue_date:
        type: string
      phone_number:
        type: string
        description: Phone number.
      email:
        type: string
      type:
        type: string
      active_areas:
        type: array
        items:
          type: integer
      org_name:
        type: string
...
  UserInfo:
    type: object
    required:
      - id
      - login
      - role
      - surname
      - name
      - patronymic
      - passport_series
      - passport_number
      - passport_authority
      - passport_issue_date
      - phone_number
      - email
      - is_readonly
      - type
      - org_name
      - is_registration_confirmed
    properties:
      id:
        type: integer
      login:
        type: string
      role:
        type: object
        properties:
          id:
            type: integer
          name:
            type: string
      surname:
        type: string
      name:
        type: string
      patronymic:
        type: string
      passport_series:
        type: string
      passport_number:
        type: string
      passport_authority:
        type: string
      passport_issue_date:
        type: string
      phone_number:
        type: string
      email:
        type: string
      is_readonly:
        type: boolean
        example: false
      type:
        type: string
      org_name:
        type: string
      is_registration_confirmed:
        type: boolean
        example: false
...
  SuccessUserUpdateResponse:
    type: object
    required:
      - data
    properties:
      data:
        $ref: "#/definitions/User"
  #Role
...